Fix build issues reported on https://gforge.inria.fr/forum/message.php?msg_id=149179

+++ src/tools/io/vpParseArgv.cpp
@@ -33,6 +33,7 @@
 #include <visp/vpParseArgv.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <string.h>
 #include <math.h>
 #include <ctype.h>
@@ -95,7 +96,7 @@ vpParseArgv::parse(int *argcPtr, const char **argv, vpArgvInfo *argTable,
 				 * than srcIndex). */
    int argc;			/* # arguments in argv still to process. */
    size_t length;			/* Number of characters in current argument. */
-   unsigned long nargs;                   /* Number of following arguments to get. */
+   unsigned long long nargs;                   /* Number of following arguments to get. */
 
 /* Macro to optionally print errors */
 #define FPRINTF if (!(flags&ARGV_NO_PRINT)) (void) fprintf
@@ -172,10 +173,10 @@ vpParseArgv::parse(int *argcPtr, const char **argv, vpArgvInfo *argTable,
       infoPtr = matchPtr;
       switch (infoPtr->type) {
       case ARGV_CONSTANT:
-	*((int *) infoPtr->dst) = (long) infoPtr->src;
+	*((long long *) infoPtr->dst) = (intptr_t) infoPtr->src;
          break;
       case ARGV_INT:
-         nargs = (unsigned long) infoPtr->src;
+         nargs = (uintptr_t) infoPtr->src;
          if (nargs<1) nargs=1;
          for (unsigned long i=0; i<nargs; i++) {
             if (argc == 0) {
@@ -197,7 +198,7 @@ vpParseArgv::parse(int *argcPtr, const char **argv, vpArgvInfo *argTable,
          }
          break;
       case ARGV_LONG:
-         nargs = (unsigned long) infoPtr->src;
+         nargs = (uintptr_t) infoPtr->src;
          if (nargs<1) nargs=1;
          for (unsigned long i=0; i<nargs; i++) {
             if (argc == 0) {
@@ -219,7 +220,7 @@ vpParseArgv::parse(int *argcPtr, const char **argv, vpArgvInfo *argTable,
          }
          break;
       case ARGV_STRING:
-         nargs = (unsigned long) infoPtr->src;
+         nargs = (uintptr_t) infoPtr->src;
          if (nargs<1) nargs=1;
          for (unsigned long i=0; i<nargs; i++) {
             if (argc == 0) {
@@ -235,7 +236,7 @@ vpParseArgv::parse(int *argcPtr, const char **argv, vpArgvInfo *argTable,
          *((int *) infoPtr->dst) = dstIndex;
          goto argsDone;
       case ARGV_FLOAT:
-         nargs = (unsigned long) infoPtr->src;
+         nargs = (uintptr_t) infoPtr->src;
          if (nargs<1) nargs=1;
          for (unsigned long i=0; i<nargs; i++) {
             if (argc == 0) {
@@ -257,7 +258,7 @@ vpParseArgv::parse(int *argcPtr, const char **argv, vpArgvInfo *argTable,
          }
          break;
       case ARGV_DOUBLE:
-         nargs = (unsigned long) infoPtr->src;
+         nargs = (uintptr_t) infoPtr->src;
          if (nargs<1) nargs=1;
          for (unsigned long i=0; i<nargs; i++) {
             if (argc == 0) {
@@ -356,7 +357,7 @@ vpParseArgv::printUsage(vpArgvInfo * argTable, int flags)
 #define NUM_SPACES 20
    static char spaces[] = "                    ";
 /*   char tmp[30]; */
-   unsigned long nargs;
+   unsigned long long nargs;
 
 /* Macro to optionally print errors */
 #define FPRINTF if (!(flags&ARGV_NO_PRINT)) (void) fprintf
@@ -403,7 +404,7 @@ vpParseArgv::printUsage(vpArgvInfo * argTable, int flags)
          switch (infoPtr->type) {
          case ARGV_INT: {
             FPRINTF(stderr, "\n\t\tDefault value:");
-            nargs = (unsigned long) infoPtr->src;
+            nargs = (uintptr_t) infoPtr->src;
             if (nargs<1) nargs=1;
             for (unsigned long j=0; j<nargs; j++) {
                FPRINTF(stderr, " %d", *(((int *) infoPtr->dst)+j));
@@ -412,7 +413,7 @@ vpParseArgv::printUsage(vpArgvInfo * argTable, int flags)
          }
          case ARGV_LONG: {
             FPRINTF(stderr, "\n\t\tDefault value:");
-            nargs = (unsigned long) infoPtr->src;
+            nargs = (uintptr_t) infoPtr->src;
             if (nargs<1) nargs=1;
             for (unsigned long j=0; j<nargs; j++) {
                FPRINTF(stderr, " %ld", *(((long *) infoPtr->dst)+j));
@@ -421,7 +422,7 @@ vpParseArgv::printUsage(vpArgvInfo * argTable, int flags)
          }
          case ARGV_FLOAT: {
             FPRINTF(stderr, "\n\t\tDefault value:");
-            nargs = (unsigned long) infoPtr->src;
+            nargs = (uintptr_t) infoPtr->src;
             if (nargs<1) nargs=1;
             for (unsigned long j=0; j<nargs; j++) {
                FPRINTF(stderr, " %f", *(((float *) infoPtr->dst)+j));
@@ -430,7 +431,7 @@ vpParseArgv::printUsage(vpArgvInfo * argTable, int flags)
          }
          case ARGV_DOUBLE: {
             FPRINTF(stderr, "\n\t\tDefault value:");
-            nargs = (unsigned long) infoPtr->src;
+            nargs = (uintptr_t) infoPtr->src;
             if (nargs<1) nargs=1;
             for (unsigned long j=0; j<nargs; j++) {
                FPRINTF(stderr, " %g", *(((double *) infoPtr->dst)+j));
@@ -440,7 +441,7 @@ vpParseArgv::printUsage(vpArgvInfo * argTable, int flags)
          case ARGV_STRING: {
             const char *string;
 
-            nargs = (unsigned long) infoPtr->src;
+            nargs = (uintptr_t) infoPtr->src;
             if (nargs<1) nargs=1;
             string = *((const char **) infoPtr->dst);
             if ((nargs==1) && (string == NULL)) break;
